+++ wetterdienst/provider/dwd/dmo/api.py
@@ -154,6 +154,7 @@ class DwdDmoValues(TimeseriesValues):
             value_name="value",
         )
         return df.with_columns(
+            pl.col("date").str.to_datetime(),
             pl.lit(station_id, dtype=pl.String).alias("station_id"),
             pl.lit(value=None, dtype=pl.Float64).alias("quality"),
         )